refactor: reorganize imports for Event and RobotProposal files for clarity
