+++ src/main/scala/io/github/srs/controller/ControllerModule.scala
@@ -1,11 +1,12 @@
 package io.github.srs.controller
 
-import scala.concurrent.duration.{ DurationInt, FiniteDuration, MILLISECONDS }
+import scala.concurrent.duration.{DurationInt, FiniteDuration, MILLISECONDS}
 import scala.language.postfixOps
-
 import cats.effect.std.Queue
-import cats.effect.{ Clock, IO }
+import cats.effect.{Clock, IO}
 import cats.syntax.all.*
+import io.github.srs.controller.message.RobotProposal
+import io.github.srs.controller.protocol.Event
 import io.github.srs.model.*
 import io.github.srs.model.SimulationConfig.SimulationStatus
 import io.github.srs.model.UpdateLogic.*
@@ -14,7 +15,7 @@ import io.github.srs.model.entity.dynamicentity.sensor.Sensor.senseAll
 import io.github.srs.model.logic.*
 import io.github.srs.utils.SimulationDefaults.debugMode
 import io.github.srs.utils.random.RNG
-import io.github.srs.utils.random.RandomDSL.{ generate, shuffle }
+import io.github.srs.utils.random.RandomDSL.{generate, shuffle}
 
 /**
  * Module that defines the controller logic for the Scala Robotics Simulator.
